Add files via upload
Adding all the required files that are needed to run my A2 notebook, as well as all of its products. 

Since the entire repository is being pulled, this should help with compilation. For example, the saved data structure for ensemble mapping has been saved so that we aren't waiting for it to come back online.